+++ src/components/App/App.scss
@@ -1,0 +1,5 @@
+@import "../Person/Person.scss";
+
+iframe {
+  display: none;
+}
