+++ src/components/Person/Person.scss
@@ -1,7 +1,3 @@
-iframe {
-  display: none;
-}
-
 .Person {
   width: fit-content;
   margin-bottom: 16px;
